refactor jwt refresh in route and restructure project
