+++ modular-website/README.md
@@ -6,7 +6,7 @@ Here are some useful links to get you started:
 
 - [Ktor Documentation](https://ktor.io/docs/home.html)
 - [Ktor GitHub page](https://github.com/ktorio/ktor)
-- The [Ktor Slack chat](https://app.slack.com/client/T09229ZC6/C0A974TJ9). You'll need to [request an invite](https://surveys.jetbrains.com/s3/kotlin-slack-sign-up) to join.
+- The [Ktor Slack chat](https://app.slack.com/client/T09229ZC6/C0A974TJ9). You'll need to [domain an invite](https://surveys.jetbrains.com/s3/kotlin-slack-sign-up) to join.
 
 ## Features
 
@@ -14,9 +14,9 @@ Here's a list of features included in this project:
 
 | Name                                                   | Description                                                    |
 | --------------------------------------------------------|---------------------------------------------------------------- |
-| [Authentication](https://start.ktor.io/p/auth)         | Provides extension point for handling the Authorization header |
+| [authentication](https://start.ktor.io/p/auth)         | Provides extension point for handling the Authorization header |
 | [Routing](https://start.ktor.io/p/routing)             | Provides a structured routing DSL                              |
-| [Authentication JWT](https://start.ktor.io/p/auth-jwt) | Handles JSON Web Token (JWT) bearer authentication scheme      |
+| [authentication JWT](https://start.ktor.io/p/auth-jwt) | Handles JSON Web Token (JWT) bearer authentication scheme      |
 
 ## Building & Running
