+++ modular-website/src/main/kotlin/authentication/AuthRoutes.kt
@@ -1,0 +1,82 @@
+package com.sundriedham.Authentication
+
+import com.sundriedham.Authentication.hashing.HashService
+import com.sundriedham.Authentication.token.JwtTokenService
+import com.sundriedham.authentication.data.user.UserRepository
+import com.sundriedham.authentication.router.AuthenticationRefreshError
+import com.sundriedham.authentication.router.AuthenticationRequestError
+import com.sundriedham.authentication.router.AuthenticationRouter
+import com.sundriedham.request.AuthenticationResponse
+import com.sundriedham.request.LoginCredentialsRequest
+import com.sundriedham.request.RefreshAuthenticationRequest
+import com.sundriedham.utils.networking.NetworkResult
+import io.ktor.http.*
+import io.ktor.server.application.*
+import io.ktor.server.auth.*
+import io.ktor.server.auth.jwt.*
+import io.ktor.server.request.*
+import io.ktor.server.response.*
+import io.ktor.server.routing.*
+
+
+suspend inline fun <reified T: Any> ApplicationCall.safeReceiveOrNull(): T? {
+    try {
+        return receive<T>()
+    } catch (e: ContentTransformationException){
+        respond(HttpStatusCode.BadRequest)
+        return null
+    }
+}
+
+fun Routing.configureAuthRoutes(
+    jwtTokenService: JwtTokenService,
+    userRepository: UserRepository,
+    hashingService: HashService
+) {
+    val router = AuthenticationRouter(
+        userRepository,
+        hashingService,
+        jwtTokenService
+    )
+
+    // "/auth/login, "/auth/refresh
+    route("auth") {
+        post("signin") {
+            val request = call.safeReceiveOrNull<LoginCredentialsRequest>() ?: return@post
+            when (val result = router.authenticate(request)) {
+                is NetworkResult.Failure<AuthenticationRequestError> -> when (result.error) {
+                    AuthenticationRequestError.PasswordInvalid ->
+                        call.respond(HttpStatusCode.Conflict, "Incorrect username or password")
+                    AuthenticationRequestError.UserNotFound ->
+                        call.respond(HttpStatusCode.Conflict, "Incorrect username or password")
+                }
+                is NetworkResult.Success<AuthenticationResponse> ->
+                    call.respond(HttpStatusCode.OK, result.response)
+            }
+        }
+
+        post("refresh") {
+            val request = call.safeReceiveOrNull<RefreshAuthenticationRequest>() ?: return@post
+            when(val result = router.authenticateRefresh(request)){
+                is NetworkResult.Failure<AuthenticationRefreshError> -> when(result.error){
+                    AuthenticationRefreshError.RefreshTokenInvalid,
+                    AuthenticationRefreshError.UserIDNotFound,
+                    AuthenticationRefreshError.UserNotFound ->
+                        call.respond(HttpStatusCode.Conflict,"Refresh token Invalid")
+                }
+                is NetworkResult.Success<AuthenticationResponse> ->
+                    call.respond(HttpStatusCode.OK, result.response)
+            }
+        }
+
+        authenticate("auth-jwt") {
+            get("/check") {
+                val principal = call.principal<JWTPrincipal>()
+                val username = principal!!.payload.getClaim("username").asString()
+                val expiresAt = principal.expiresAt?.time?.minus(System.currentTimeMillis())
+                call.respondText("Hello, $username! Token is expired at $expiresAt ms.")
+            }
+        }
+    }
+}
+
