+++ modular-website/src/main/kotlin/authentication/data/user/UserRepository.kt
@@ -1,4 +1,4 @@
-package com.sundriedham.data.user
+package com.sundriedham.authentication.data.user
 
 interface UserRepository {
     suspend fun getUserByUserid(id: Identifier<User>): User?
