+++ modular-website/src/main/kotlin/utils/networking/NetworkResult.kt
@@ -1,0 +1,6 @@
+package com.sundriedham.utils.networking
+
+sealed class NetworkResult<out T, out Error> {
+    class Success<T>(val response: T): NetworkResult<T, Nothing>()
+    class Failure<Error>(val error: Error): NetworkResult<Nothing, Error>()
+}
