+++ modular-website/src/main/resources/test.conf
@@ -1,0 +1,16 @@
+ktor {
+    deployment {
+        port = 8080
+        port = ${?PORT}
+    }
+    application {
+        modules = [ com.sundriedham.ApplicationKt.module ]
+    }
+}
+jwt {
+    issuer = "https://jwt-provider-domain/"
+    domain = "http://0.0.0.0:8080"
+    jwtAudience = "jwt-audience"
+    refreshTokenAudience = "sundriedham-refresh"
+    realm = "ktor sample app"
+}
